Add caesar and xor ciphers, change to BigInteger

+++ src/main/java/chat/ClientInfo.java
@@ -2,17 +2,18 @@ package chat;
 
 import lombok.Data;
 
+import java.math.BigInteger;
 import java.net.Socket;
 
 /**
- * Created by Beata on 2016-10-19.
+ * Created by Beata Kalis on 2016-10-19.
  */
 @Data
 public class ClientInfo {
 
     private Socket socket;
 
-    private Long secret; // shared secret between client and server
+    private BigInteger secret; // shared secret between client and server
 
     private String encryption;
